Fixed color support issues
There were some issues with how color support was handled. They have
been fixed by working with character support to an increased degree.
This means that the first two characters after \003 are checked to see
if they are digits. If the first or both are, they are put into an int
which then becomes the color code. The default is color reset, which
simply means that the sequence will be ignored.